Removed dummy file
Removed the dummy feature rule that was used to call static island at 0,0. Also changed island feature rules so no static island is created.

Some players had problems with the pack not loading due to world.x fault. The dummy was the fix for that fault, but is sensitive to order of operation for different device OS. Removed this to improve reliability across devices.